requirements.txt
flask==2.0.1
openai-whisper==20231117
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0
webvtt-py==0.4.6
werkzeug==2.0.1
aeidon==1.12
gunicorn==20.1.0
